102 editor squad (#106)
* Add: Basic Scene for Squad Editor

#102

+ 스쿼드 에디터용 씬 기본 구성

* Feat: nullable enable

#102

~ nullable 오브젝트 #nullable enable 선언

* Add: Select Monster

#103

+ 배치할 칸 클릭 후 몬스터 선택창 오픈

* Add: Adding monster

#102

+ 몬스터 클릭해서 배치

* Add: Delete monster

#102

+ 배치한 몬스터 삭제 기능

* Add: Save & Load

#102

+ 저장, 불러오기, 다른 이름으로 저장 기능 구현
+ 엔트리 위치 관련 변수 전부 float -> int

+++ SlimeGG/Assets/Scripts/Battles/Manager/BattleManager.cs
@@ -391,7 +391,7 @@ public class BattleManager : MonoBehaviour
         inventoryManager.setEntriable(true);
         foreach (ContentController enemy in enemyEntry)
         {
-            float[] pos = enemy.monsterLiveStat.saveStat.entryPos;
+            int[] pos = enemy.monsterLiveStat.saveStat.entryPos;
             fieldController.enemyEntryGenerator.transform.GetChild((int)pos[0] + ((int)pos[1] * SettingVariables.Battle.entrySizeMax[0])).GetComponent<EntrySlotController>().installMonster(enemy);
         }
         fieldController.allyEntryGenerator.sendCellsFunction((Transform cell) => cell.GetComponent<EntrySlotController>().isAlly = true);
