+++ SlimeGG/Assets/Scripts/Objects/Save/MonsterSaveStat.cs
@@ -7,5 +7,5 @@ public class MonsterSaveStat
     public List<ElementStat> exp = new List<ElementStat>();
     public string id;
     public string speicie;
-    public float[] entryPos;
+    public int[] entryPos;
 }
